Migrated UpdateNotes to Coroutines
